Move Data module folder to store folder
