+++ src/ADempiere/modules/persistence/store/Data/mutations.ts
@@ -4,7 +4,7 @@ import {
   BusinessDataState,
   IPrivateAccessDataExtended,
   IRecordSelectionData
-} from '../PersistenceType'
+} from '../../PersistenceType'
 
 type BusinessDataMutationTree = MutationTree<BusinessDataState>
